Created primenumber.h file for lab 6
